move handleInvite to viewers list
