+++ apps/web/src/components/ParticipantWindow/ParticipantWindow.tsx
@@ -1,12 +1,9 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
 import clsx from 'clsx';
 import { useAppState } from '../../state';
 import ParticipantWindowHeader from './ParticipantWindowHeader/ParticipantWindowHeader';
 import ViewersList from '../ViewersList/ViewersList';
-import useVideoContext from '../../hooks/useVideoContext/useVideoContext';
-import { sendSpeakerInvite } from '../../state/api/api';
-import { useEnqueueSnackbar } from '../../hooks/useSnackbar/useSnackbar';
 import SpeakersList from '../SpeakersList/SpeakersList';
 
 const useStyles = makeStyles((theme: Theme) =>
@@ -35,26 +32,12 @@ const useStyles = makeStyles((theme: Theme) =>
 export default function ParticipantWindow() {
   const classes = useStyles();
   const { appState } = useAppState();
-  const { room } = useVideoContext();
-  const enqueueSnackbar = useEnqueueSnackbar();
-
-  const handleInvite = useCallback(
-    (raisedHand: string) => {
-      sendSpeakerInvite(raisedHand, room!.sid);
-      enqueueSnackbar({
-        headline: 'Invite Sent',
-        message: `You invited ${raisedHand} to be a speaker. They will now be able to share audio and video.`,
-        variant: 'info',
-      });
-    },
-    [room, enqueueSnackbar]
-  );
 
   return (
     <aside className={clsx(classes.participantWindowContainer, { [classes.hide]: !appState.isParticipantWindowOpen })}>
       <ParticipantWindowHeader />
       <SpeakersList />
-      <ViewersList handleInvite={handleInvite} />
+      <ViewersList />
     </aside>
   );
 }
