+++ apps/web/src/components/ViewersList/ViewersList.tsx
@@ -1,9 +1,12 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import { Typography } from '@material-ui/core';
 import { useAppState } from '../../state';
+import { useEnqueueSnackbar } from '../../hooks/useSnackbar/useSnackbar';
 import { useRaisedHandsMap } from '../../hooks/useRaisedHandsMap/useRaisedHandsMap';
+import useVideoContext from '../../hooks/useVideoContext/useVideoContext';
 import { useViewersMap } from '../../hooks/useViewersMap/useViewersMap';
+import { sendSpeakerInvite } from '../../state/api/api';
 import { RaisedHand } from '../ParticipantWindow/RaisedHand/RaisedHand';
 
 const useStyles = makeStyles({
@@ -20,18 +23,28 @@ const useStyles = makeStyles({
   },
 });
 
-interface ViewersListProps {
-  handleInvite: (handleInviteIdentity: string) => void;
-}
-
-export default function ViewersList({ handleInvite }: ViewersListProps) {
+export default function ViewersList() {
   const { appState } = useAppState();
+  const { room } = useVideoContext();
+  const enqueueSnackbar = useEnqueueSnackbar();
   const raisedHands = useRaisedHandsMap();
   const viewers = useViewersMap();
   const viewersWithoutRaisedHands = viewers.filter(viewer => !raisedHands.includes(viewer));
   const viewerCount = viewersWithoutRaisedHands.length + raisedHands.length;
 
   const classes = useStyles();
+
+  const handleInvite = useCallback(
+    (raisedHand: string) => {
+      sendSpeakerInvite(raisedHand, room!.sid);
+      enqueueSnackbar({
+        headline: 'Invite Sent',
+        message: `You invited ${raisedHand} to be a speaker. They will now be able to share audio and video.`,
+        variant: 'info',
+      });
+    },
+    [room, enqueueSnackbar]
+  );
 
   return (
     <div className={classes.viewersContainer}>
@@ -42,12 +55,13 @@ export default function ViewersList({ handleInvite }: ViewersListProps) {
           name={raisedHand}
           handleInvite={handleInvite}
           isHost={appState.participantType === 'host'}
+          isLocalViewer={appState.participantName === raisedHand}
         />
       ))}
 
       {viewersWithoutRaisedHands.map(viewer => (
         <Typography key={viewer} variant="body1">
-          {viewer}
+          {appState.participantName === viewer ? `${viewer} (You)` : viewer}
         </Typography>
       ))}
     </div>
